test: Add unit tests + fix model typo
- Add dependencies: MockK, Turbine, and Coroutines Test
- Implement unit tests for `EstimateDistanceUseCase`, `DiscoveredDevice`, `DiscoveryViewModel`, and `BleScanner`
- Rename `DisocveredDevice` to `DiscoveredDevice` and update timestamp type to Double
- Expose `calculateSmoothedRssi` in `BleScanner` for testing
- Add `@RequiresPermission` annotations to `DiscoveryScreen` actions

+++ app/src/main/java/com/fyp/crowdlink/domain/model/DiscoveredDevice.kt
@@ -3,6 +3,6 @@ package com.fyp.crowdlink.domain.model
 data class DiscoveredDevice(
     val deviceId: String,
     val rssi: Int,
-    val timestamp: Long = System.currentTimeMillis(),
+    val timestamp: Double = System.currentTimeMillis().toDouble(),
     val estimatedDistance: Double = -1.0
 )
